Fix serialization issues.
Previously, I observed different results when running watersheds on the cluster vs on my local machine. I added Serializer for all broadcasts or replaced the objects to be broadcast with objects that kryo knows how to serialize.
My best bet is that the culprit was
```java
TObjectLongHashMap< HashWrapper< long[] > > offsetMap = ...
sc.broadcast( offsetMap );
```